remove old hmac, new stubggen
