--- stdlib/2.7/hmac.pyi
@@ -1,13 +1,0 @@
-# Stubs for hmac (Python 2)
-
-from typing import Any, AnyStr
-
-class HMAC:
-    def update(self, msg: str) -> None: ...
-    def digest(self) -> str: ...
-    def hexdigest(self) -> str: ...
-    def copy(self) -> HMAC: ...
-
-def new(key: str, msg: str = ..., digestmod: Any = ...) -> HMAC: ...
-
-def compare_digest(a: AnyStr, b: AnyStr) -> bool: ...
