+++ stdlib/2and3/hmac.pyi
@@ -1,0 +1,27 @@
+# Stubs for hmac (Python 3.4)
+#
+# NOTE: This dynamically typed stub was automatically generated by stubgen.
+
+from typing import Any, Optional
+from _operator import _compare_digest as compare_digest
+
+trans_5C = ...  # type: Any
+trans_36 = ...  # type: Any
+digest_size = ...  # type: Optional[Any]
+
+class HMAC:
+    blocksize = ...  # type: int
+    digest_cons = ...  # type: Any
+    outer = ...  # type: Any
+    inner = ...  # type: Any
+    digest_size = ...  # type: Any
+    block_size = ...  # type: Any
+    def __init__(self, key, msg: Optional[Any] = ..., digestmod: Optional[Any] = ...) -> None: ...
+    @property
+    def name(self): ...
+    def update(self, msg): ...
+    def copy(self): ...
+    def digest(self): ...
+    def hexdigest(self): ...
+
+def new(key, msg: Optional[Any] = ..., digestmod: Optional[Any] = ...): ...
