--- stdlib/3/hmac.pyi
@@ -1,27 +1,0 @@
-# Stubs for hmac (Python 3.5)
-#
-# NOTE: This dynamically typed stub was automatically generated by stubgen.
-
-from typing import Any
-from _operator import _compare_digest as compare_digest
-
-trans_5C = ... # type: Any
-trans_36 = ... # type: Any
-digest_size = ... # type: Any
-
-class HMAC:
-    blocksize = ... # type: Any
-    digest_cons = ... # type: Any
-    outer = ... # type: Any
-    inner = ... # type: Any
-    digest_size = ... # type: Any
-    block_size = ... # type: Any
-    def __init__(self, key, msg=None, digestmod=None): ...
-    @property
-    def name(self): ...
-    def update(self, msg): ...
-    def copy(self): ...
-    def digest(self): ...
-    def hexdigest(self): ...
-
-def new(key, msg=None, digestmod=None): ...
